Removed shell as compile-time dependency from neo and a NeoService can now instead start a ShellServer purely via reflections.
The shell got a dependency on neo and doesn't need to to everything
via reflections anymore. It can, however, still be used as a remote client
w/o the neo/jta jars on the classpath (java -jar shell.jar) for
connecting to an already running neo service.

This dependency hierarchy feels better and more right than the
other way around.


git-svn-id: https://svn.neo4j.org/components/shell/trunk@3304 0b971d98-bb2f-0410-8247-b05b2b5feb2a